refactor: improve skin data organization
