+++ src/shared/data/skins/types.ts
@@ -1,0 +1,19 @@
+import { images } from "shared/assets";
+
+export interface SnakeSkin {
+	readonly id: string;
+	readonly price: number;
+	readonly size: Vector2;
+	readonly tint: readonly Color3[];
+	readonly texture: readonly string[];
+	readonly primary?: Color3;
+	readonly secondary?: Color3;
+}
+
+export const defaultSnakeSkin: SnakeSkin = {
+	id: "default",
+	price: 0,
+	size: new Vector2(512, 512),
+	tint: [Color3.fromRGB(255, 255, 255)],
+	texture: [images.skins.snake_main],
+};
