+++ src/shared/store/saves/save-types.ts
@@ -1,5 +1,5 @@
 import { t } from "@rbxts/t";
-import { defaultSnakeSkins } from "shared/data/skins";
+import { baseSnakeSkins } from "shared/data/skins";
 
 import { PlayerSave } from "./save-slice";
 
@@ -7,7 +7,7 @@ export const RANDOM_SKIN = "__random__";
 
 export const defaultPlayerSave: PlayerSave = {
 	balance: 0,
-	skins: [RANDOM_SKIN, ...defaultSnakeSkins.map((skin) => skin.id)],
+	skins: [RANDOM_SKIN, ...baseSnakeSkins.map((skin) => skin.id)],
 	skin: RANDOM_SKIN,
 };
